# Decrease the initial nap for wait_sshable from 8 to 7 seconds