Over time, we updated the Cloud Hypervisor version, migrated some hosts to the new storage layer (ubiblk), and improved the performance of respirate.

We have a significant number of wait_sshable runs that are less than 9 seconds in production. Therefore, we can safely reduce the initial nap to 7 seconds.
<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Reduce initial nap time in `wait_sshable` from 8 to 7 seconds in `nexus.rb` based on performance data.
> 
>   - **Behavior**:
>     - Decrease initial nap time in `wait_sshable` from 8 to 7 seconds in `nexus.rb`.
>     - Update test in `nexus_spec.rb` to expect a 7-second nap instead of 8.
>   - **Rationale**:
>     - Performance improvements and data indicate `wait_sshable` often completes in under 9 seconds, allowing for a reduced nap time.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=ubicloud%2Fubicloud&utm_source=github&utm_medium=referral)<sup> for 87e582a84c1362b7a2a3f526b4a81667f84265f6. You can [customize](https://app.ellipsis.dev/ubicloud/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->